refactor: make searchbox more customizeable
