+++ instantsearch-compose/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="search_box_hint">Search…</string>
+    <string name="alg_is_compose_search_box_hint">Search…</string>
 </resources>
